:alembic: ci: Rework CI to build Sphinx docs and push to GitHub Pages
This commit fixes some of the logic for building and pushing automatic
updates to the GitHub Pages site when source content is changed in the
repository. Specific details of note:

* Fix documentation build paths, do not push to `docs/` folder
* Trigger deploy job on new commits added to `main` branch
* Standardized docs project name "prokuranepal.github.io"

This commit alone will not work. It also needs a `$GITHUB_TOKEN` set as
a secret variable from the Travis CI web interface for the deploy step
to work. Without the token, there is no way for Travis to push the
generated HTML to GitHub Pages, because it will not have access.

See also:

	https://github.com/jwflory/sphinx-docs-opinionated-quickstart

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>